refactored chest loot for Emi, implemented and updated config, added debug switch and logger integrated into the config. Added resource-pack-driven entity fixer system for the mob loot.

+++ src/main/java/fzzyhmstrs/emi_loot/EMILoot.java
@@ -14,17 +14,21 @@ import net.minecraft.util.math.random.Random;
 import net.minecraft.util.registry.Registry;
 
 import java.io.*;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.Arrays;
+import java.util.logging.*;
 
 public class EMILoot implements ModInitializer {
 
-    public static boolean DEBUG = false;
     public static String MOD_ID = "emi_loot";
+    public static final Logger LOGGER = getLogger();
     public static Random emiLootRandom = new LocalRandom(System.currentTimeMillis());
     public static LootTableParser parser = new LootTableParser();
-    private static Gson gson = new GsonBuilder().setPrettyPrinting().create();
+    private static final Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
     public static EmiLootConfig config = readOrCreate();
+    public static boolean DEBUG = config.debugMode;
     public static Enchantment RANDOM = new Enchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.TRIDENT, EquipmentSlot.values()){
         @Override
         public boolean isAvailableForEnchantedBookOffer() {
@@ -41,39 +45,68 @@ public class EMILoot implements ModInitializer {
         parser.registerServer();
         Registry.register(Registry.ENCHANTMENT,new Identifier(MOD_ID,"random"),RANDOM);
     }
+
+    public static void debugText(String text){
+
+    }
     
     private static EmiLootConfig readOrCreate(){
         File dir = FabricLoader.getInstance().getConfigDir().toFile();
         
         if (!dir.exists() && !dir.mkdirs()) {
-            System.out.println("EMI Loot could not find or create config directory, using default configs");
+            LOGGER.severe("EMI Loot could not find or create config directory, using default configs");
             return new EmiLootConfig();
         }
+        String f_old_name = "EmiLootConfig.json";
+        String f_name = "EmiLootConfig_v1.json";
         
-        File f = new File(dir,"EmiLootConfig.json");
+        File f_old = new File(dir,f_old_name);
         
         try{
-            if (f.exists()){
-                return gson.fromJson(new InputStreamReader(new FileInputStream(f)),EmiLootConfig.class);
-            } else if (!f.createNewFile()){
-                throw new UnsupportedOperationException("couldn't generate config file");
+            if (f_old.exists()){
+                EmiLootConfigOld oldConfig = gson.fromJson(new InputStreamReader(new FileInputStream(f_old)),EmiLootConfigOld.class);
+                EmiLootConfig newConfig = oldConfig.generateNewConfig();
+                File f = new File(dir,f_name);
+                if (f.exists()){
+                    f_old.delete();
+                    return gson.fromJson(new InputStreamReader(new FileInputStream(f)),EmiLootConfig.class);
+                } else if (!f.createNewFile()){
+                    LOGGER.severe("Failed to create new config file, using old config with new defaults.");
+                } else {
+                    f_old.delete();
+                    FileWriter fw = new FileWriter(f);
+                    String json = gson.toJson(newConfig);
+                    if (EMILoot.DEBUG) EMILoot.LOGGER.info(json);
+                    fw.write(json);
+                    fw.close();
+                }
+                return newConfig;
             } else {
-                FileWriter fw = new FileWriter(f);
-                EmiLootConfig emc = new EmiLootConfig();
-                String json = gson.toJson(emc);
-                System.out.println(json);
-                fw.write(json);
-                fw.close();
-                return emc;
+                File f = new File(dir,f_name);
+                if (f.exists()) {
+                    return gson.fromJson(new InputStreamReader(new FileInputStream(f)), EmiLootConfig.class);
+                } else if (!f.createNewFile()) {
+                    throw new UnsupportedOperationException("couldn't generate config file");
+                } else {
+                    FileWriter fw = new FileWriter(f);
+                    EmiLootConfig emc = new EmiLootConfig();
+                    String json = gson.toJson(emc);
+                    if (EMILoot.DEBUG) EMILoot.LOGGER.info(json);
+                    fw.write(json);
+                    fw.close();
+                    return emc;
+                }
             }
         } catch(Exception e){
-            System.out.println("Emi Loot failed to create or read it's config file!");
-            System.out.println(Arrays.toString(e.getStackTrace()));
+            LOGGER.severe("Emi Loot failed to create or read it's config file!");
+            LOGGER.severe(Arrays.toString(e.getStackTrace()));
             return new EmiLootConfig();
         }
     }
     
     public static class EmiLootConfig{
+        public boolean debugMode = false;
+
         public boolean parseChestLoot = true;
         
         public boolean parseBlockLoot = true;
@@ -81,5 +114,50 @@ public class EMILoot implements ModInitializer {
         public boolean parseMobLoot = true;
     
         public boolean parseGameplayLoot = true;
+
+        public boolean chestLootCompact = true;
+
+        public boolean chestLootAlwaysStackSame = false;
+    }
+
+    public static class EmiLootConfigOld{
+        public boolean parseChestLoot = true;
+
+        public boolean parseBlockLoot = true;
+
+        public boolean parseMobLoot = true;
+
+        public boolean parseGameplayLoot = true;
+
+        public EmiLootConfig generateNewConfig(){
+            EmiLootConfig newConfig = new EmiLootConfig();
+            newConfig.parseChestLoot = this.parseChestLoot;
+            newConfig.parseBlockLoot = this.parseBlockLoot;
+            newConfig.parseMobLoot = this.parseMobLoot;
+            newConfig.parseGameplayLoot = this.parseGameplayLoot;
+            return newConfig;
+        }
+    }
+
+    private static Logger getLogger(){
+        Logger logger = Logger.getLogger(MOD_ID);
+        ConsoleHandler handler = new ConsoleHandler();
+        handler.setFormatter(new LogFormat());
+        handler.setLevel(Level.ALL);
+        logger.setLevel(Level.ALL);
+        logger.addHandler(handler);
+        logger.setUseParentHandlers(false);
+        return logger;
+    }
+
+    private static class LogFormat extends Formatter{
+        @Override
+        public String format(LogRecord record) {
+            ZonedDateTime zdt = ZonedDateTime.ofInstant(record.getInstant(), ZoneId.systemDefault());
+            String a = "[" + zdt.getHour() + ":" + zdt.getMinute() + ":" + zdt.getSecond() + "] ";
+            String b = "[EMI Loot/" + record.getLevel().getLocalizedName() + "] [" + record.getSourceClassName() + "]: ";
+            String c = record.getMessage() + "\n";
+            return a + b + c;
+        }
     }
 }
